requirements/requirements.txt
pony
asyncio-nats-client
ujson
